backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic-settings==2.5.2
httpx==0.27.2
python-dotenv==1.0.1
python-multipart==0.0.12
email-validator==2.1.1